Code reformat on source tree.

+++ src/main/java/macrobase/analysis/summary/itemset/FPGrowthEmerging.java
@@ -25,22 +25,22 @@ public class FPGrowthEmerging {
 
 
     @SuppressWarnings("unused")
-	private static final Logger log = LoggerFactory.getLogger(FPGrowthEmerging.class);
+    private static final Logger log = LoggerFactory.getLogger(FPGrowthEmerging.class);
 
     private List<ItemsetResult> getSingletonItemsets(List<DatumWithScore> inliers,
                                                      List<DatumWithScore> outliers,
                                                      double minSupport,
                                                      double minRatio,
                                                      DatumEncoder encoder) {
-        int supportCountRequired = (int)(outliers.size()*minSupport);
+        int supportCountRequired = (int) (outliers.size() * minSupport);
 
         List<ItemsetResult> ret = new ArrayList<>();
 
         Map<Integer, Double> inlierCounts = new ExactCount().count(inliers).getCounts();
         Map<Integer, Double> outlierCounts = new ExactCount().count(outliers).getCounts();
 
-        for(Map.Entry<Integer, Double> outlierCount : outlierCounts.entrySet()) {
-            if(outlierCount.getValue() < supportCountRequired) {
+        for (Map.Entry<Integer, Double> outlierCount : outlierCounts.entrySet()) {
+            if (outlierCount.getValue() < supportCountRequired) {
                 continue;
             }
 
@@ -48,14 +48,14 @@ public class FPGrowthEmerging {
 
             double ratio;
 
-            if(inlierCount != null) {
-                ratio = (outlierCount.getValue()/ outliers.size()) /
-                         (inlierCount/ inliers.size());
+            if (inlierCount != null) {
+                ratio = (outlierCount.getValue() / outliers.size()) /
+                        (inlierCount / inliers.size());
             } else {
                 ratio = Double.POSITIVE_INFINITY;
             }
 
-            if(ratio > minRatio) {
+            if (ratio > minRatio) {
                 ret.add(new ItemsetResult(outlierCount.getValue() / outliers.size(),
                                           outlierCount.getValue(),
                                           ratio,
@@ -72,7 +72,7 @@ public class FPGrowthEmerging {
                                                                  double minRatio,
                                                                  // would prefer not to pass this in, but easier for now...
                                                                  DatumEncoder encoder) {
-        if(inliers.get(0).getDatum().getAttributes().size() == 1) {
+        if (inliers.get(0).getDatum().getAttributes().size() == 1) {
             return getSingletonItemsets(inliers, outliers, minSupport, minRatio, encoder);
         }
 
@@ -86,28 +86,28 @@ public class FPGrowthEmerging {
 
         Map<Integer, Double> supportedOutlierCounts = new HashMap<>();
 
-        int supportCountRequired = (int)(outliers.size()*minSupport);
+        int supportCountRequired = (int) (outliers.size() * minSupport);
 
-        for(DatumWithScore d : outliers) {
+        for (DatumWithScore d : outliers) {
             Set<Integer> txn = null;
 
-            for(int i : d.getDatum().getAttributes()) {
+            for (int i : d.getDatum().getAttributes()) {
                 double outlierCount = outlierCounts.get(i);
-                if(outlierCount >= supportCountRequired) {
+                if (outlierCount >= supportCountRequired) {
                     Number inlierCount = inlierCounts.get(i);
 
                     double outlierInlierRatio;
-                    if(inlierCount == null || inlierCount.doubleValue() == 0) {
+                    if (inlierCount == null || inlierCount.doubleValue() == 0) {
                         outlierInlierRatio = Double.POSITIVE_INFINITY;
                     } else {
-                        outlierInlierRatio = (outlierCount/outliers.size())/(inlierCount.doubleValue()/inliers.size());
+                        outlierInlierRatio = (outlierCount / outliers.size()) / (inlierCount.doubleValue() / inliers.size());
                     }
-                    if(outlierInlierRatio > minRatio) {
-                        if(txn == null) {
+                    if (outlierInlierRatio > minRatio) {
+                        if (txn == null) {
                             txn = new HashSet<>();
                         }
 
-                        if(!supportedOutlierCounts.containsKey(i)) {
+                        if (!supportedOutlierCounts.containsKey(i)) {
                             supportedOutlierCounts.put(i, outlierCount);
                         }
 
@@ -116,7 +116,7 @@ public class FPGrowthEmerging {
                 }
             }
 
-            if(txn != null) {
+            if (txn != null) {
                 outlierTransactions.add(txn);
             }
         }
@@ -125,7 +125,7 @@ public class FPGrowthEmerging {
         context = outlierFPGrowth.time();
         FPGrowth fpg = new FPGrowth();
         List<ItemsetWithCount> iwc = fpg.getItemsetsWithSupportCount(outlierTransactions, supportedOutlierCounts,
-                                                                     outliers.size()*minSupport);
+                                                                     outliers.size() * minSupport);
         context.stop();
 
         context = inlierRatio.time();
@@ -140,9 +140,9 @@ public class FPGrowthEmerging {
 
         Set<Integer> prevSet = null;
         Double prevCount = -1.;
-        for(ItemsetWithCount i : iwc) {
-            if(i.getCount() == prevCount) {
-                if(prevSet != null && Sets.difference(i.getItems(), prevSet).size() == 0) {
+        for (ItemsetWithCount i : iwc) {
+            if (i.getCount() == prevCount) {
+                if (prevSet != null && Sets.difference(i.getItems(), prevSet).size() == 0) {
                     continue;
                 }
             }
@@ -152,17 +152,17 @@ public class FPGrowthEmerging {
             prevSet = i.getItems();
 
 
-            if(i.getItems().size() == 1) {
+            if (i.getItems().size() == 1) {
                 double ratio = 0;
                 Number inlierCount = inlierCounts.get(i.getItems().iterator().next());
 
-                if(inlierCount != null && inlierCount.doubleValue() > 0) {
-                    ratio = ((double)i.getCount()/outliers.size())/((double)inlierCount/inliers.size());
+                if (inlierCount != null && inlierCount.doubleValue() > 0) {
+                    ratio = (i.getCount() / outliers.size()) / ((double) inlierCount / inliers.size());
                 } else {
                     ratio = Double.POSITIVE_INFINITY;
                 }
-                
-                ret.add(new ItemsetResult(i.getCount()/(double)outliers.size(),
+
+                ret.add(new ItemsetResult(i.getCount() / (double) outliers.size(),
                                           i.getCount(),
                                           ratio,
                                           encoder.getColsFromAttrSet(i.getItems())));
@@ -183,20 +183,20 @@ public class FPGrowthEmerging {
                                                                            ratioItemsToCheck,
                                                                            ratioSetsToCheck);
 
-        assert(matchingInlierCounts.size() == ratioSetsToCheck.size());
-        for(int i = 0; i < matchingInlierCounts.size(); ++i) {
+        assert (matchingInlierCounts.size() == ratioSetsToCheck.size());
+        for (int i = 0; i < matchingInlierCounts.size(); ++i) {
             ItemsetWithCount ic = matchingInlierCounts.get(i);
             ItemsetWithCount oc = ratioSetsToCheck.get(i);
 
             double ratio;
-            if(ic.getCount() > 0) {
-                ratio = (oc.getCount()/outliers.size())/((double)ic.getCount()/inliers.size());
+            if (ic.getCount() > 0) {
+                ratio = (oc.getCount() / outliers.size()) / (ic.getCount() / inliers.size());
             } else {
                 ratio = Double.POSITIVE_INFINITY;
             }
 
-            if(ratio >= minRatio) {
-                ret.add(new ItemsetResult(oc.getCount()/(double)outliers.size(),
+            if (ratio >= minRatio) {
+                ret.add(new ItemsetResult(oc.getCount() / (double) outliers.size(),
                                           oc.getCount(),
                                           ratio,
                                           encoder.getColsFromAttrSet(oc.getItems())));
